init: Relax user check for chrome to run kiosk UI with user chronos
The factory kiosk UI is running with user chronos. If system settings
is configured to only allow certain users to log in, the Chrome
terminates itself due to chronos is not in the allow list, thus causing
no CDP endpoint for goofy to interact with.

Add the option `--oobe-skip-postlogin` to skip user check such that even
if system settings are configured to only allow certain users to log in,
the chrome still allows chronos to run the kiosk UI.

BUG=b:271193237
TEST=Follow b/271193237#comment3 to turn on 'Restrict sign-in' option,
     install factory toolkit and reboot. After reboot, can enter factory
     kiosk UI, and http://localhost:9222/json/list endpoint is
     accessible.

Change-Id: I0bdca0981e75e491f69ef63f36918c9ecf0d7c3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4909368
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>