hwid/service: add expect_probe_value field to skip match
As there is an implicit type conversion in the expect field (e.g.
"hex"), we should not remove the field if we don't want to match them in
the probe config.  This CL adds a field expect_probe_value to skip the
probe value but keep it an optional field instead in the probe config.

BUG=b:301916718
TEST=./deploy/cros_hwid_service.sh test

Change-Id: I53ed1779dea3a34aeeff8f3a9ac027e79ef3531e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4910682
Reviewed-by: Kevin Lin <kevinptt@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>