hwid/service: allow field names to be qual-specific
For probe info requiring probe attributes from qualifications, the probe
value converter will not receive a valid probe info if the component in
HWID DB does not embed the qid in the component name (e.g. storage
subcomp of a storage assembly or a qualification is not yet created).
This CL adds `qual_specific` attribute in `ConvertedValueSpec` class and
`is_qual_probe_info` in arguments of `converter*.Match` methods so that
the converter can optionally skip probe attributes which are
qual-specific.

BUG=b:304711720
TEST=./deploy/cros_hwid_service.sh test

Change-Id: I21e10184fe8b43a32ed9fba72281c35b12dffc9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4929598
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>