sh: device_id.sh: generate device_id by gsc device id
After crrev/c/4936800 is landed, we can get GSC device id by:
```
gsctool -a -K dev_ids
```
on devices using cr50 or ti50.

The GSC device id is set in GUC factory when the GSC chip
is manufactured and can never be changed.
So we change to use GSC device id to generate device_id because
it is guaranteed to be fully unique across all GSC chips.

BUG=b:305609896, b:302054623
TEST=(1)Manually checks the output of script on DUT.
     (2)Runs `RebootSequence` test on DUT.

Change-Id: I82cf5931ab45fabe2a6eb5773c5f9918f258d662
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4951334
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: LIU KUAN-FU <iamjeffliu@google.com>
Tested-by: LIU KUAN-FU <iamjeffliu@google.com>