hwid/service: add PadEncodingBits change unit
As the existing AddEncodingCombination change unit does not cover a case
where an existing pattern adds bits of existing (but new to the pattern)
encoded fields.  This CL adds a new change unit type PadEncodingBits and
it should always be voted as approved or the resulting HWID DB might be
invalid.

BUG=b:305907572
TEST=./deploy/cros_hwid_service.sh test

Change-Id: I1f05f87263dea698491c7282398deb8e2d535535
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/4950416
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>